update check samplesheet
